refactor to make everything much prettier
