+++ frontend/src/metabase-types/api/bookmark.ts
@@ -1,10 +1,11 @@
 export type BookmarkType = "card" | "collection" | "dashboard";
+export type BookmarkId = string;
 
 export interface Bookmark {
   authority_level?: string;
   card_id?: string;
   display?: string;
-  id: string;
+  id: BookmarkId;
   item_id: number;
   name: string;
   type: BookmarkType;
