+++ frontend/src/metabase-types/api/collection.ts
@@ -7,7 +7,7 @@ import type { TableId } from "./table";
 
 export type RegularCollectionId = number;
 
-export type CollectionId = RegularCollectionId | "root" | "personal";
+export type CollectionId = RegularCollectionId | "root" | "personal" | "users";
 
 export type CollectionContentModel = "card" | "dataset";
 
@@ -97,5 +97,7 @@ export interface CollectionItem {
 export interface CollectionListQuery {
   archived?: boolean;
   "exclude-other-user-collections"?: boolean;
+  "exclude-archived"?: boolean;
   namespace?: string;
+  tree?: boolean;
 }
