+++ frontend/src/metabase/collections/components/CollectionContent/CollectionContent.styled.tsx
