+++ frontend/src/metabase/collections/components/CollectionContent/index.ts
@@ -1,0 +1,1 @@
+export * from "./CollectionContent";
