+++ frontend/src/metabase/collections/components/CollectionLanding/CollectionLanding.tsx
@@ -1,6 +1,7 @@
 import type { ReactNode } from "react";
 import { extractCollectionId } from "metabase/lib/urls";
-import { CollectionContent } from "../../containers/CollectionContent";
+import { isNotNull } from "metabase/lib/types";
+import { CollectionContent } from "../CollectionContent";
 
 export interface CollectionLandingProps {
   params: CollectionLandingParams;
@@ -16,6 +17,10 @@ const CollectionLanding = ({
   children,
 }: CollectionLandingProps) => {
   const collectionId = extractCollectionId(slug);
+
+  if (!isNotNull(collectionId)) {
+    return null;
+  }
 
   return (
     <>
